[minor] resharper suggested changes

+++ Src/Metrics/Utils/EWMA.cs
@@ -71,7 +71,7 @@ namespace Metrics.Utils
 
         public double GetRate(TimeUnit rateUnit)
         {
-            return rate.Get() * (double)rateUnit.ToNanoseconds(1L);
+            return rate.Get() * rateUnit.ToNanoseconds(1L);
         }
 
         public void Reset()
